api/requirements.txt
fastapi
pandas
scikit-learn
pydantic
uvicorn[standart]
gunicorn
python-dotenv
os